test: add benchmark and unit tests to cover optimized resolve solution
